chore: move util to separate file
