+++ lib/triage.js
@@ -1,25 +1,8 @@
 'use strict';
 
 const calcChainProfit = require('./calcChainProfit');
+const { getConnectingAsset, getMultiplier } = require('./util');
 
-// TODO: Get this fee from the market data (this is binance fee 0.1%);
-const fee = 0.001;
-
-const getBase = symbol => symbol.market.base;
-const getQuote = symbol => symbol.market.quote;
-const inversePrice = n => 1 / n;
-
-const getMultiplier = (symbol, target, ticker) => getBase(symbol) === target ? ticker.info.bidPrice : inversePrice(ticker.info.bidPrice);
-
-const getConnectingAsset = (symbol, target) => {
-    const base = getBase(symbol);
-    
-    if (base === target) {
-       return getQuote(symbol);
-    }
-
-    return base;
-}
 
 module.exports = async function calculateTriage(exchange, target, symbol1, symbol2, symbol3) {
 
@@ -43,12 +26,13 @@ module.exports = async function calculateTriage(exchange, target, symbol1, symbo
         const c = getMultiplier(symbol3, connectingAsset2, ticker3);
 
         // Calculate profits
+        let fee = 0.001
         const difference = calcChainProfit({ a, b, c, fee });
-        
+
         // Return findings
         const prefix = [target, connectingAsset1, connectingAsset2, target].join(' > ');
 
-        console.log(prefix + ': ' + difference + '%');      
+        console.log(prefix + ': ' + difference + '%');
     });
 
 }
